backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.14.11
sqlparse==0.5.5
uritemplate==4.2.0
gunicorn
psycopg2-binary
djangorestframework-simplejwt
